Move UserInfo to common
