+++ catalogue/webapp/components/ConfirmItemRequest/ConfirmItemRequest.tsx
@@ -10,7 +10,7 @@ import { PhysicalItem, Work } from '@weco/common/model/catalogue';
 import { classNames, font } from '@weco/common/utils/classnames';
 import LL from '@weco/common/views/components/styled/LL';
 import { withAppPathPrefix } from '@weco/identity/src/utility/app-path-prefix';
-import { UserInfo } from '@weco/identity/src/frontend/MyAccount/UserInfoContext/UserInfo.interface';
+import { UserInfo } from '@weco/common/model/user';
 
 const Header = styled(Space).attrs({
   v: { size: 'm', properties: ['margin-bottom'] },
